+++ src/IdentityServer4.Contrib.Caching.Redis/Configuration/RedisCacheGrantStoreConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace IdentityServer4.Contrib.Caching.Redis.Configuration
+{
+    public class RedisCacheGrantStoreConfiguration
+    {
+        public string CachingKeyPrefix { get; set; }
+    }
+}