+++ src/IdentityServer4.Contrib.Caching.Redis/Extensions/StringExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace IdentityServer4.Contrib.Caching.Abstractions.Extensions
+namespace IdentityServer4.Contrib.Caching.Redis.Extensions
 {
-    public static class StringExtensions
+    internal static class StringExtensions
     {
         public static void EnsureParamter(this string value, string parameterName)
         {