+++ test/IdentityServer4.Contrib.Caching.Redis.Tests/Misc/ServiceProviderFixture.cs
@@ -1,18 +1,36 @@
 ﻿using System;
-using IdentityServer4.Contrib.Caching.Abstractions.Extensions;
+using Elders.RedLock;
+using IdentityServer4.Contrib.Caching.Redis.Configuration;
 using IdentityServer4.Contrib.Caching.Redis.Extensions;
 using IdentityServer4.Contrib.Caching.Redis.Stores;
+using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.StackExchangeRedis;
 using Microsoft.Extensions.DependencyInjection;
+using Moq;
+
+// ReSharper disable MemberCanBeMadeStatic.Global
 
 namespace IdentityServer4.Contrib.Caching.Redis.Tests.Misc
 {
     public class ServiceProviderFixture
     {
+        public Mock<IDistributedCache> CreateDistributedCacheMock() => new Mock<IDistributedCache>();
+        
+        public Mock<IRedisLockManager> CreateRedisLockManagerMock() => new Mock<IRedisLockManager>();
+
+        public IServiceProvider BuildMockServiceProvider(IDistributedCache cache, IRedisLockManager redisLockManager)
+            => new ServiceCollection()
+                .AddIdentityServerBuilder()
+                .AddPersistedGrantStore<RedisCacheGrantStore>()
+                .Services
+                .AddSingleton(cache)
+                .AddSingleton(redisLockManager)
+                .Configure<RedisCacheGrantStoreConfiguration>(options => options.CachingKeyPrefix = "TEST_PREFIX")
+                .BuildServiceProvider();
+        
         public IServiceProvider BuildDefaultServiceProvider(RedisCacheOptions options)
             => new ServiceCollection()
                 .AddIdentityServerBuilder()
-                .AddDistributedCacheGrantStore<RedisCacheGrantStoreService>()
                 .AddDistributedRedisCache(options.Configuration, options.InstanceName)
                 .Services
                 .BuildServiceProvider();
@@ -20,8 +38,8 @@ namespace IdentityServer4.Contrib.Caching.Redis.Tests.Misc
         public IServiceProvider BuildDefaultServiceProvider(Action<RedisCacheOptions> options)
             => new ServiceCollection()
                 .AddIdentityServerBuilder()
-                .AddDistributedCacheGrantStore<RedisCacheGrantStoreService>()
-                .AddDistributedRedisCache(options)
+                .AddDistributedRedisCache(options,
+                    cacheOptions => cacheOptions.CachingKeyPrefix = "_IdentityServer_Redis_Cache_Grant_Store_")
                 .Services
                 .BuildServiceProvider();
     }